Refactor: Phase 1 Domain-Driven Design (DDD) layout
- Created new core directories (`core/domains`, `core/system`, `core/common`)
- Relocated files from `spec/domains` to `core/domains`
- Moved `core/exceptions.py` to `core/common/exceptions.py`
- Moved `core/rebuild.py` to `core/system/rebuild.py`
- Removed all inline imports globally to enforce strict standard
- Verified code changes and structure updates locally with mypy and ruff.
