+++ src/coreason_manifest/cli/main.py
@@ -16,8 +16,8 @@ import typer
 from rich import print as rprint
 
 from coreason_manifest.adapters.system.dynamic_loader import load_flow_from_file
+from coreason_manifest.core.domains.scientific_vis import HierarchicalBlueprint
 from coreason_manifest.core.workflow.topologies.sci_vis_flow import get_sota_scivis_topology
-from coreason_manifest.spec.domains.scientific_vis import HierarchicalBlueprint
 from coreason_manifest.toolkit.exporter import render_agent_card
 from coreason_manifest.toolkit.validator import validate_flow
 from coreason_manifest.toolkit.visualizer import export_html_diagram, to_mermaid
