+++ src/coreason_manifest/core/workflow/nodes/human.py
@@ -5,8 +5,8 @@ from typing import Any, Literal
 from pydantic import Field, model_validator
 
 from coreason_manifest.core.common.base import CoreasonModel
+from coreason_manifest.core.common.exceptions import ManifestError, ManifestErrorCode
 from coreason_manifest.core.common.presentation import AdaptiveUIContract, RenderStrategy
-from coreason_manifest.core.exceptions import ManifestError, ManifestErrorCode
 from coreason_manifest.core.oversight.resilience import EscalationStrategy
 from coreason_manifest.core.primitives.registry import register_node
 from coreason_manifest.core.primitives.types import VariableID
